Refactor: Report Directory
