+++ green/src/main/java/com/greengrim/green/core/report/controller/RegisterReportController.java
@@ -2,7 +2,7 @@ package com.greengrim.green.core.report.controller;
 
 import com.greengrim.green.common.oauth.auth.CurrentMember;
 import com.greengrim.green.core.member.Member;
-import com.greengrim.green.core.report.ReportType;
+import com.greengrim.green.core.report.entity.ReportType;
 import com.greengrim.green.core.report.dto.ReportRequestDto.RegisterReport;
 import com.greengrim.green.core.report.service.RegisterReportService;
 import io.swagger.v3.oas.annotations.Operation;
