+++ green/src/main/java/com/greengrim/green/core/report/entity/Report.java
@@ -1,4 +1,4 @@
-package com.greengrim.green.core.report;
+package com.greengrim.green.core.report.entity;
 
 import com.greengrim.green.common.entity.BaseTime;
 import com.greengrim.green.core.member.Member;
