+++ green/src/main/java/com/greengrim/green/core/report/entity/ReportType.java
@@ -1,4 +1,4 @@
-package com.greengrim.green.core.report;
+package com.greengrim.green.core.report.entity;
 
 public enum ReportType {
     MEMBER, CHALLENGE, CERTIFICATION, NFT
