+++ green/src/main/java/com/greengrim/green/core/report/repository/ReportRepository.java
@@ -1,6 +1,6 @@
 package com.greengrim.green.core.report.repository;
 
-import com.greengrim.green.core.report.Report;
+import com.greengrim.green.core.report.entity.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReportRepository extends JpaRepository<Report, Long> {
