+++ green/src/main/java/com/greengrim/green/core/report/service/RegisterReportService.java
@@ -17,8 +17,8 @@ import com.greengrim.green.core.member.Member;
 import com.greengrim.green.core.member.repository.MemberRepository;
 import com.greengrim.green.core.nft.Nft;
 import com.greengrim.green.core.nft.repository.NftRepository;
-import com.greengrim.green.core.report.Report;
-import com.greengrim.green.core.report.ReportType;
+import com.greengrim.green.core.report.entity.Report;
+import com.greengrim.green.core.report.entity.ReportType;
 import com.greengrim.green.core.report.dto.ReportRequestDto.RegisterReport;
 import com.greengrim.green.core.report.repository.ReportRepository;
 import jakarta.transaction.Transactional;
